+++ CHANGES.md
@@ -6,10 +6,11 @@
 
 * `App::configure()` allow to offload app configuration to different methods
 
+* Added `URLPath` option for logger
+
 * Added `ServiceRequest::app_data()`, returns `Data<T>`
 
 * Added `ServiceFromRequest::app_data()`, returns `Data<T>`
-
 
 ### Changed